+++ base/src/main/java/org/killbill/billing/platform/config/KillbillConfigSourceChain.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2014-2019 Groupon, Inc
+ * Copyright 2014-2019 The Billing Project, LLC
+ *
+ * The Billing Project licenses this file to you under the Apache License, version 2.0
+ * (the "License"); you may not use this file except in compliance with the
+ * License.  You may obtain a copy of the License at:
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.killbill.billing.platform.config;
+
+import java.util.List;
+
+import org.killbill.billing.platform.api.KillbillConfigSource;
+
+public class KillbillConfigSourceChain implements KillbillConfigSource {
+
+    private final List<KillbillConfigSource> sources;
+
+    public KillbillConfigSourceChain(List<KillbillConfigSource> sources) {
+        this.sources = sources;
+    }
+
+    @Override
+    public String getString(final String propertyName) {
+        return sources.stream()
+                      .filter(source -> source.hasProperty(propertyName))
+                      .map(source -> source.getString(propertyName))
+                      .findFirst()
+                      .orElse(null);
+    }
+
+    @Override
+    public boolean hasProperty(final String propertyName) {
+        return sources.stream().anyMatch(source -> source.hasProperty(propertyName));
+    }
+}
