+++ base/src/test/java/org/killbill/billing/platform/config/TestEnvKillbillConfigSource.java
@@ -1,6 +1,6 @@
 /*
- * Copyright 2014 Groupon, Inc
- * Copyright 2014 The Billing Project, LLC
+ * Copyright 2014-2019 Groupon, Inc
+ * Copyright 2014-2019 The Billing Project, LLC
  *
  * The Billing Project licenses this file to you under the Apache License, version 2.0
  * (the "License"); you may not use this file except in compliance with the
@@ -15,12 +15,19 @@
  * under the License.
  */
 
-package org.killbill.billing.platform.api;
+package org.killbill.billing.platform.config;
 
-public interface KillbillConfigSource {
+import java.io.IOException;
+import java.net.URISyntaxException;
 
-    String getString(String propertyName);
+import org.testng.Assert;
+import org.testng.annotations.Test;
 
-    boolean hasProperty(String propertyName);
+public class TestEnvKillbillConfigSource {
 
+    @Test(groups = "fast")
+    public void testNameMapping() throws IOException, URISyntaxException {
+        String name = EnvKillbillConfigSource.propertyNameToEnvName("killbill.test.prop");
+        Assert.assertEquals(name, "KILLBILL_TEST_PROP");
+    }
 }
